+++ src/FunFair.CodeAnalysis.Tests/ProhibitedMethodInvocationsDiagnosticsAnalyzerTests.cs
@@ -109,5 +109,172 @@ namespace FunFair.CodeAnalysis.Tests
 
             return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.Assert}, expected);
         }
+
+        [Fact]
+        public Task ShouldRaiseErrorForAddOrUpdateAsync()
+        {
+            const string test = @"
+     using NonBlocking;
+
+     namespace ConsoleApplication1
+     {
+         class TypeName
+         {
+             int Update(int x, int y){
+                return x+y+1;
+            }
+
+             void Test()
+             {
+                 ConcurrentDictionary<int,int> dictionary = new ConcurrentDictionary<int,int>();
+                 dictionary.AddOrUpdate(key: 1, addValueFactory: (x) => x + 1, updateValueFactory: this.Update);
+             }
+         }
+     }";
+            DiagnosticResult expected = new()
+                                        {
+                                            Id = "FFS0032",
+                                            Message = @"Don't use any of the built in AddOrUpdate methods, instead FunFair.Common.Extensions.ConcurrentDictionaryExtensions.AddOrUpdate can be used",
+                                            Severity = DiagnosticSeverity.Error,
+                                            Locations = new[] {new DiagnosticResultLocation(path: "Test0.cs", line: 15, column: 18)}
+                                        };
+
+            return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.NonBlockingConcurrentDictionary}, expected);
+        }
+
+        [Fact]
+        public Task ShouldRaiseErrorForAddOrUpdateWithOutAddValueFactoryAsync()
+        {
+            const string test = @"
+     using NonBlocking;
+
+     namespace ConsoleApplication1
+     {
+         class TypeName
+         {
+            void Test()
+             {
+                 ConcurrentDictionary<int,int> dictionary = new ConcurrentDictionary<int,int>();
+                 dictionary.AddOrUpdate(key: 1, addValue: 1, updateValueFactory: (x, y) => x + y);
+             }
+         }
+     }";
+            DiagnosticResult expected = new()
+                                        {
+                                            Id = "FFS0032",
+                                            Message = @"Don't use any of the built in AddOrUpdate methods, instead FunFair.Common.Extensions.ConcurrentDictionaryExtensions.AddOrUpdate can be used",
+                                            Severity = DiagnosticSeverity.Error,
+                                            Locations = new[] {new DiagnosticResultLocation(path: "Test0.cs", line: 11, column: 18)}
+                                        };
+
+            return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.NonBlockingConcurrentDictionary}, expected);
+        }
+
+        [Fact]
+        public Task ShouldRaiseErrorForGetOrAddAsync()
+        {
+            const string test = @"
+     using NonBlocking;
+
+     namespace ConsoleApplication1
+     {
+         class TypeName
+         {
+            void Test()
+             {
+                 ConcurrentDictionary<int,int> dictionary = new ConcurrentDictionary<int,int>();
+                 dictionary.GetOrAdd(key: 1, valueFactory: i => i);
+             }
+         }
+     }";
+            DiagnosticResult expected = new()
+                                        {
+                                            Id = "FFS0033",
+                                            Message = @"Don't use any of the built in GetOrAdd methods, instead FunFair.Common.Extensions.ConcurrentDictionaryExtensions.GetOrAdd can be used",
+                                            Severity = DiagnosticSeverity.Error,
+                                            Locations = new[] {new DiagnosticResultLocation(path: "Test0.cs", line: 11, column: 18)}
+                                        };
+
+            return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.NonBlockingConcurrentDictionary}, expected);
+        }
+
+        [Fact]
+        public Task ShouldAllowAddOrUpdateAsync()
+        {
+            const string test = @"
+     using NonBlocking;
+
+     namespace ConsoleApplication1
+     {
+         class TypeName
+         {
+             int Update(int x, int y){
+                return x+y+1;
+            }
+
+            int Add(int x){
+                return x+1;
+            }
+
+             void Test()
+             {
+                 ConcurrentDictionary<int,int> dictionary = new ConcurrentDictionary<int,int>();
+                 dictionary.AddOrUpdate(key: 1, addValueFactory: this.Add, updateValueFactory: this.Update);
+             }
+         }
+     }";
+
+            return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.NonBlockingConcurrentDictionary});
+        }
+
+        [Fact]
+        public Task ShouldAllowAddOrUpdateWithOutAddValueFactoryAsync()
+        {
+            const string test = @"
+     using NonBlocking;
+
+     namespace ConsoleApplication1
+     {
+         class TypeName
+         {
+             int Update(int x, int y){
+                return x+y+1;
+            }
+
+             void Test()
+             {
+                 ConcurrentDictionary<int,int> dictionary = new ConcurrentDictionary<int,int>();
+                 dictionary.AddOrUpdate(key: 1, addValue: 1, updateValueFactory: this.Update);
+             }
+         }
+     }";
+
+            return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.NonBlockingConcurrentDictionary});
+        }
+
+        [Fact]
+        public Task ShouldAllowGetOrAddAsync()
+        {
+            const string test = @"
+     using NonBlocking;
+
+     namespace ConsoleApplication1
+     {
+         class TypeName
+         {
+             int ValueFactory(int x){
+                return x+1;
+            }
+
+             void Test()
+             {
+                 ConcurrentDictionary<int,int> dictionary = new ConcurrentDictionary<int,int>();
+                 dictionary.GetOrAdd(key: 1, valueFactory: this.ValueFactory);
+             }
+         }
+     }";
+
+            return this.VerifyCSharpDiagnosticAsync(source: test, new[] {WellKnownMetadataReferences.NonBlockingConcurrentDictionary});
+        }
     }
 }